+++ Java_pool/Day05/ex01/ChatEx01/src/main/java/edu/school21/chat/repositories/MessagesRepository.java
@@ -5,5 +5,5 @@ import edu.school21.chat.models.Message;
 import java.util.Optional;
 
 public interface MessagesRepository {
-    Optional<Message> findById(int id);
+    Optional<Message> findById(Long id);
 }
